ARTEMIS-3261 Updating logic to use only replaceable records on compacting verification
(cherry picked from commit 0edf599adca9365993579a96d717ed2f3c0d0af9)

downstream: ENTMQBR-4831

Note: This commit was actually originally developed as part of 2.16.0.jbossorg-x and then cherry-picked into upstream and rebased there.

+++ artemis-cli/src/main/java/org/apache/activemq/artemis/cli/commands/tools/journal/CompactJournal.java
@@ -59,11 +59,11 @@ public final class CompactJournal extends LockAbstract {
                                final int poolFiles,
                                final int fileSize,
                                final IOCriticalErrorListener listener,
-                               int... replaceableRecords) throws Exception {
+                               byte... replaceableRecords) throws Exception {
       NIOSequentialFileFactory nio = new NIOSequentialFileFactory(directory, listener, 1);
 
       JournalImpl journal = new JournalImpl(fileSize, minFiles, poolFiles, 0, 0, nio, journalPrefix, journalSuffix, 1);
-      for (int i : replaceableRecords) {
+      for (byte i : replaceableRecords) {
          journal.replaceableRecord(i);
       }
       journal.setRemoveExtraFilesOnLoad(true);
